requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
pytickersymbols
#scipy
pydantic-settings
deta==1.2.0